+++ fuzz/fuzz_targets/rangemap_inclusive_gaps.rs
@@ -31,12 +31,13 @@ impl Arbitrary for Input {
     fn arbitrary(u: &mut Unstructured) -> arbitrary::Result<Self> {
         Ok(Self {
             ops: u.arbitrary()?,
-            // Larger margins than that are too
+            // Larger margins than these are too
             // far away from boundary conditions to be interesting.
             // ("Oh, the fools! If only they'd built it with 6,001 hulls." -- Philip J. Fry)
             //
             // NOTE: Not using `int_in_range` because of <https://github.com/rust-fuzz/arbitrary/issues/106>.
-            outer_range: *u.choose(&[0, 1, 2, 3])?..=*u.choose(&[252, 253, 254, 255])?,
+            outer_range: *u.choose(&[0, 1, 2, 3, 100, 101, 102, 103])?
+                ..=*u.choose(&[100, 101, 102, 103, 252, 253, 254, 255])?,
         })
     }
 }
@@ -66,9 +67,20 @@ fuzz_target!(|input: Input| {
             u8::max(*range.start(), *outer_range.start())
                 ..=u8::min(*range.end(), *outer_range.end())
         })
+        .filter(|range| {
+            // Reject anything that is now empty after being truncated.
+            !range.is_empty()
+        })
         .collect();
     keys.extend(gaps.into_iter());
     keys.sort_by_key(|key| *key.start());
+
+    if outer_range.is_empty() {
+        // There should be no gaps or keys returned if the outer range is empty,
+        // because empty ranges cover no values.
+        assert!(keys.is_empty());
+        return;
+    }
 
     // Gaps and keys combined should span whole outer range.
     assert_eq!(keys.first().unwrap().start(), outer_range.start());